Readme file updated

+++ tasks/models.py
@@ -17,8 +17,8 @@ class Task(models.Model):
     performers: str = models.CharField(max_length=255, blank=True)  # Список исполнителей в виде строки
     created_at: timezone.datetime = models.DateTimeField(default=timezone.now)  # Дата и время создания
     status: str = models.CharField(max_length=50, choices=STATUS_CHOICES, default="assigned")  # Статус 
-    planned_effort: float = models.DecimalField(default=0.0)  # Плановое время выполнения
-    actual_effort: float = models.DecimalField(default=0.0)  # Фактическое время
+    planned_effort: float = models.FloatField(default=0.0)  # Плановое время выполнения
+    actual_effort: float = models.FloatField(default=0.0)  # Фактическое время
     completed_at: Optional[timezone.datetime] = models.DateTimeField(null=True, blank=True)  # Дата завершения (опционально)
 
     parent: Optional["Task"] = models.ForeignKey(
